Add blockValue to EngineGetPayloadResultV2
Ported code over from https://github.com/hyperledger/besu/pull/4833
